requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Pillow==10.2.0
black==24.3.0
flake8==7.0.0
pytest==8.1.1
pytest-django==4.8.0
coverage==7.4.4 